requirements.txt
# Core Framework
typer>=0.9.0
rich>=13.0.0
asyncpg>=0.28.0

# LLM & HTTP
openai>=1.0.0
aiohttp>=3.8.0
httpx>=0.24.0

# Data & Storage
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pyyaml>=6.0.0

# Utilities
colorama>=0.4.6
click>=8.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
